+++ src/main/java/jpa/domain/common/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package jpa.domain;
+package jpa.domain.common;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
