+++ src/main/java/jpa/domain/line/Line.java
@@ -1,4 +1,7 @@
-package jpa.domain;
+package jpa.domain.line;
+
+import jpa.domain.common.BaseTimeEntity;
+import jpa.domain.stationline.StationLine;
 
 import javax.persistence.*;
 import java.util.ArrayList;
