+++ src/main/java/jpa/domain/line/LineColor.java
@@ -1,4 +1,4 @@
-package jpa.domain;
+package jpa.domain.line;
 
 public enum LineColor {
     BLUE, GREEN, ORANGE
