+++ src/main/java/jpa/domain/line/LineRepository.java
@@ -1,5 +1,6 @@
-package jpa.domain;
+package jpa.domain.line;
 
+import jpa.domain.line.Line;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LineRepository extends JpaRepository<Line, Long> {
