+++ src/main/java/jpa/domain/station/StationRepository.java
@@ -1,5 +1,6 @@
-package jpa.domain;
+package jpa.domain.station;
 
+import jpa.domain.station.Station;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
