+++ src/main/java/jpa/domain/stationline/StationLineRepository.java
@@ -1,5 +1,8 @@
-package jpa.domain;
+package jpa.domain.stationline;
 
+import jpa.domain.line.Line;
+import jpa.domain.station.Station;
+import jpa.domain.stationline.StationLine;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StationLineRepository extends JpaRepository<StationLine, Long> {
