+++ src/test/java/jpa/domain/MemberRepositoryTest.java
@@ -1,5 +1,7 @@
 package jpa.domain;
 
+import jpa.domain.member.Member;
+import jpa.domain.member.MemberRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
@@ -16,7 +18,7 @@ class MemberRepositoryTest {
 
     @Test
     void save() {
-        Member member = new Member(25, "1234");
+        Member member = new Member(25, "asd@gamil.com", "1234");
         Member actual = memberRepository.save(member);
 
         assertThat(actual.getId()).isNotNull();
@@ -25,7 +27,7 @@ class MemberRepositoryTest {
 
     @Test
     void findByAge() {
-        memberRepository.save(new Member(24, "1111"));
+        memberRepository.save(new Member(24, "asd@gamil.com", "1111"));
         memberRepository.save(new Member(24, "kj@xxx.com", "222"));
 
         List<Member> members = memberRepository.findAllByAge(24);
@@ -45,7 +47,7 @@ class MemberRepositoryTest {
 
     @Test
     void delete() {
-        Member member = memberRepository.save(new Member(24, "1111"));
+        Member member = memberRepository.save(new Member(24, "woo@xxx.com", "1111"));
 
         memberRepository.delete(member);
         Optional<Member> actual = memberRepository.findById(member.getId());
